Fix additional format warning
Apparently, this only happens on 64-bit platforms.

+++ src/interfaces/ecpg/ecpglib/error.c
@@ -336,7 +336,7 @@ ecpg_raise_backend(int line, PGresult *result, PGconn *conn, int compat)
 
 	/* %.*s is safe here as long as sqlstate is all-ASCII */
 	ecpg_log("raising sqlstate %.*s (sqlcode %ld): %s\n",
-			 sizeof(sqlca->sqlstate), sqlca->sqlstate, sqlca->sqlcode, sqlca->sqlerrm.sqlerrmc);
+			 (int) sizeof(sqlca->sqlstate), sqlca->sqlstate, sqlca->sqlcode, sqlca->sqlerrm.sqlerrmc);
 
 	/* free all memory we have allocated for the user */
 	ECPGfree_auto_mem();
